chore: rename DataMocks to data-mocks
